refactor(channels): RFC 0035 PR 4 — consolidate member + history handlers into one file
Relocate handleGetMembershipHistory + membershipIntervalsToResponse from
channel_membership_handlers.go into channel_member_handlers.go (next to
handleUpdateChannelMember); delete the emptied file and rename its test to
channel_member_history_test.go. Behaviour-preserving: the handler is already
wired in channel_routes.go (PR #674) — this only groups the member-facing
handlers into one aptly-named file.

docs(channels): complete the REST route table in docs/guides/channels.md by
adding the four shipped-but-undocumented endpoints, each verified against
channel_routes.go: GET .../activity (RFC 0048), PATCH .../members/{participant_id}
(RFC 0050), and GET + PATCH .../config (RFC 0050).

docs(rfcs): update 0035-pr-plan.md to reflect the handler-file rename.
Regenerate FILEMAP.md; minor sqlite_membership_intervals.go and
channel_handlers_test.go touch-ups.
