+++ FILEMAP.md
@@ -1,13 +1,13 @@
 # Repository File Map
 
-> Auto-generated by `scripts/generate_filemap.py` — 2026-06-18  
-> **1544** tracked files
+> Auto-generated by `scripts/generate_filemap.py` — 2026-06-19  
+> **1543** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 373 |
+| `internal/` | 372 |
 | `docs/` | 360 |
 | `tests/` | 325 |
 | `agents/` | 234 |
@@ -30,7 +30,7 @@
 |-----------|-------|
 | `.py` | 576 |
 | `.md` | 413 |
-| `.go` | 387 |
+| `.go` | 386 |
 | `.js` | 41 |
 | `.yaml` | 37 |
 | `.rs` | 29 |
@@ -1122,9 +1122,8 @@
 │   │   ├── channel_handlers_review_test.go
 │   │   ├── channel_handlers_test.go
 │   │   ├── channel_member_handlers.go
+│   │   ├── channel_member_history_test.go
 │   │   ├── channel_member_update_test.go
-│   │   ├── channel_membership_handlers.go
-│   │   ├── channel_membership_handlers_test.go
 │   │   ├── channel_mention_lift.go
 │   │   ├── channel_mention_lift_test.go
 │   │   ├── channel_publish_async_test.go
